+++ ChangeLog.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 
+- add arangodb.httpCookieSpec
 - host handling (issue #241)
 - logging extended hostresolver